Listados

+++ dam/cawas/Controller/SliderController.py
@@ -268,16 +268,17 @@ class SliderController(object):
 
             # LOGO
             if (request.FILES.has_key('logo')):
-                if request.FILES['logo'].name != '':
-                    # TRATAMIENTO DE IMAGEN LOGO
-                    vimg.logo = request.FILES['logo']
-                    extension = os.path.splitext(vimg.logo.name)[1]
-                    varchivo = pathfileslogo.value + vimg.name + extension
-                    vimg.logo.name = varchivo
-                    varchivo_server = base_dir.value + varchivo
-                    if os.path.isfile(varchivo_server):
-                        os.remove(varchivo_server)
-                        # /LOGO
+                if request.FILES['logo'].name is None:
+                    if request.FILES['logo'].name != '':
+                        # TRATAMIENTO DE IMAGEN LOGO
+                        vimg.logo = request.FILES['logo']
+                        extension = os.path.splitext(vimg.logo.name)[1]
+                        varchivo = pathfileslogo.value + vimg.name + extension
+                        vimg.logo.name = varchivo
+                        varchivo_server = base_dir.value + varchivo
+                        if os.path.isfile(varchivo_server):
+                            os.remove(varchivo_server)
+                            # /LOGO
 
 
             vimg.save()
